refactor(update_project): extract socials to child form WIP
